requirements.txt
transformers>=4.28.0
torch
Pillow
requests
